refactored out to IO wrapper class
Moved the nonblocking IO wrapper to a class to be more explicit and
reusable.

Added minitest gem and test/unit test directory. Here I didn't want to
use RSpec to test whether non-blocking IO is functional on the current
platform, because it didn't feel like a specification test. My feeling
was more of a xunit-style setup/teardown.

Added test to reasonably determine that IO is nonblocking by reading
smaller windows of a subprocess's IO.