renamed extensions to rdf_repository
The "resource_repository" name implied the usage of the RDF data and not
the true purpose which is to read/write RDF data.

These extensions allow RDF data to be used from different sources. The
usage of the RDF repositories (e.g. Namespaces, Annotations) apply the
expectations on the data.