support JSON abstraction in core
The JSON abstraction was moved to core since it is a shared dependency
between the JSON and JGF translator extensions.

This abstraction allows streaming support where implemented. Currently
only in the Oj adapter. This is sort-of a lightweight MultiJson
implementation since it does not provide streaming API due to lack of
general support.